refactor: universe settings redux slice
